fix(engine): resolve CodeQL findings for REST/SDK interop (BEN-84)
